organized things better. inline edit/delete now working
