--- app/templates/layout.css
@@ -1,6 +1,0 @@
-/* layout.css Style */
-.upload-drop-zone {
-  height: 200px;
-  border-width: 2px;
-  margin-bottom: 20px;
-}
