+++ config.py
@@ -1,0 +1,4 @@
+ALLOWED_EXTENSIONS = set(['pdf'])
+NON_PROFITS_FOLDER = "nonprofits/"
+AUTH_SERVER = "localhost"
+AUTH_PORT = 13370
