+++ docs/adding_templates.txt
